food_recipe_system/requirements.txt
virtualenv==20.13.3
Django==3.1.7
asgiref==3.4.1
pytz==2021.1
gunicorn
whitenoise==5.2.0
django-jsonfield==1.4.1